refactor: Extract train_epoch to train_step.py (Sequence 4A)
EXTRACTION:
- Moved train_epoch (614 lines) from loop.py to train_step.py
- Updated loop.py to import train_epoch from train_step
- All dependencies and imports moved correctly

RESULTS:
- loop.py: 1746 → 1133 lines (613 lines removed)
- train_step.py: 319 lines (clean module)
- All tests pass

Part of Sequence 4 (REFACTORING_PLAN_V4.md Phase 2)
