+++ src/brain_brr/train/loop.py
@@ -25,7 +25,6 @@ import torch
 import torch.multiprocessing as mp
 import torch.nn as nn
 import torch.nn.functional as tnf
-from torch.amp import GradScaler, autocast
 from torch.optim import AdamW, Optimizer
 from torch.optim.lr_scheduler import LambdaLR, LRScheduler
 from torch.utils.data import DataLoader, WeightedRandomSampler
@@ -50,7 +49,8 @@ from src.brain_brr.config.schemas import (
 )
 from src.brain_brr.models import SeizureDetector
 from src.brain_brr.train.checkpoint import load_checkpoint, save_checkpoint
-from src.brain_brr.train.train_utils import get_memory_stats, set_seed, worker_init_fn
+from src.brain_brr.train.train_step import train_epoch
+from src.brain_brr.train.train_utils import set_seed, worker_init_fn
 from src.brain_brr.train.wandb_integration import WandBLogger
 from src.brain_brr.utils.env import env
 
@@ -334,619 +334,6 @@ def create_scheduler(
 
 
 # ============================================================================
-# Training epoch (Single Responsibility)
-# ============================================================================
-
-
-def train_epoch(
-    model: nn.Module,
-    dataloader: DataLoader,
-    optimizer: Optimizer,
-    device: str = "cpu",
-    use_amp: bool = False,
-    gradient_clip: float = 1.0,
-    scheduler: LRScheduler | None = None,
-    global_step: int = 0,
-    *,
-    loss_mode: str = "bce",
-    focal_alpha: float = 0.25,
-    focal_gamma: float = 2.0,
-    return_step: bool = False,
-    checkpoint_dir: Path | None = None,
-    epoch_index: int | None = None,
-    mid_epoch_minutes: float | None = None,
-    mid_epoch_keep: int = 3,
-    warmup_schedule: WarmupScheduleConfig | None = None,
-) -> float | tuple[float, int]:
-    """Train for one epoch.
-
-    Args:
-        model: SeizureDetector model
-        dataloader: Training DataLoader
-        optimizer: Optimizer instance
-        device: Device to train on
-        use_amp: Use automatic mixed precision
-        gradient_clip: Max gradient norm
-        scheduler: Optional LR scheduler (per-iteration)
-        global_step: Global step counter for scheduler
-        return_step: If True, return (loss, global_step). If False, return just loss.
-        warmup_schedule: Optional warmup schedule configuration for gradient stabilization
-
-    Returns:
-        Average training loss (default) or tuple of (loss, global_step) if return_step=True
-    """
-    import time
-
-    model.train()
-    device_obj = torch.device(device)
-    # Only construct GradScaler when actually using CUDA AMP
-    scaler = GradScaler(enabled=(use_amp and device == "cuda"))
-
-    supports_training_state = hasattr(model, "set_training_state")
-    if supports_training_state:
-        try:
-            model.set_training_state(global_step, warmup_schedule)
-        except TypeError:
-            supports_training_state = False
-
-    # Heartbeat timer for Modal visibility
-    last_heartbeat = time.time()
-    heartbeat_interval = 120  # 2 minutes (reduced from 5 for faster hang detection)
-    last_mid_save = time.time()
-    mid_interval_s = (
-        None if mid_epoch_minutes is None else float(max(0.0, mid_epoch_minutes)) * 60.0
-    )
-
-    # Calculate class weights from dataset sample (not just first batch!)
-    # Sample a significant portion to get accurate statistics
-    logger.info("\n" + "=" * 60)
-    logger.info("[INIT] DATASET STATISTICS")
-    logger.info("=" * 60)
-
-    dataset = dataloader.dataset
-    dataset_len = len(dataset)  # type: ignore[arg-type]
-
-    # Skip expensive sampling in smoke test mode
-    is_smoke_test = env.smoke_test()
-    if is_smoke_test:
-        logger.info("[SMOKE TEST MODE] Skipping dataset sampling - using default pos_weight=1.0")
-        pos_weight_val = 1.0
-        pos_ratio = 0.5  # Assume balanced for smoke test
-
-    # CRITICAL FIX: BalancedSeizureDataset already knows its exact seizure ratio!
-    # No need to sample 1000 windows (which takes 2+ hours on Modal)
-    else:
-        from src.brain_brr.data.datasets import BalancedSeizureDataset
-
-        if isinstance(dataset, BalancedSeizureDataset):
-            # Use the pre-computed ratio from manifest statistics
-            pos_ratio = dataset.seizure_ratio
-            logger.info("[DATASET] Using BalancedSeizureDataset known distribution")
-            logger.info(f"[DATASET] Seizure ratio: {100 * pos_ratio:.1f}% (from manifest)")
-        else:
-            # Fallback: sample windows for regular datasets
-            sample_size = min(100, dataset_len)  # Reduced from 1000 for speed
-            sample_indices = torch.randperm(dataset_len)[:sample_size]
-
-            pos_count = 0
-            total_samples = 0
-
-            logger.info(f"[DATASET] Sampling {sample_size} windows to estimate distribution...")
-            for idx in sample_indices:
-                batch = dataset[idx.item()]
-                label = batch["label"]
-                if (label > 0).any():
-                    pos_count += 1
-                total_samples += 1
-
-            pos_ratio = pos_count / total_samples if total_samples > 0 else 1e-8
-            logger.info(f"[DATASET] Sampled {sample_size} windows")
-            logger.info(
-                f"[DATASET] Windows with seizures: {pos_count}/{sample_size} ({100 * pos_ratio:.2f}%)"
-            )
-
-        # Use sqrt scaling for extreme imbalance (prevents explosion)
-        if pos_ratio > 0 and pos_ratio < 0.5:
-            pos_weight_val = math.sqrt((1 - pos_ratio) / pos_ratio)
-        else:
-            pos_weight_val = 1.0
-
-        logger.info(f"[DATASET] Using pos_weight: {pos_weight_val:.2f} (sqrt scaling)")
-    logger.info("=" * 60 + "\n")
-
-    # Validate dataset has seizures
-    if pos_ratio < 0.001:  # Less than 0.1% seizures
-        logger.critical("\n" + "!" * 60)
-        logger.critical(f"[CRITICAL] Dataset has only {100 * pos_ratio:.4f}% seizures!")
-        logger.critical("[CRITICAL] Model will likely collapse to all-negative predictions.")
-        logger.critical("[CRITICAL] Increase BGB_LIMIT_FILES or use different data split.")
-        logger.critical("!" * 60 + "\n")
-
-    # Get first batch for preflight check
-    first_batch = next(iter(dataloader))
-
-    # AMP-safe, numerically stable loss on logits (build per-element loss fn)
-    # (We keep model outputs as probabilities elsewhere for tests/inference)
-    pos_weight_t = torch.as_tensor(pos_weight_val, device=device_obj, dtype=torch.float32)
-    use_focal = (loss_mode or "bce").lower() == "focal"
-
-    current_step_ref: dict[str, int] = {"value": global_step}
-
-    if use_focal:
-        focal = FocalLoss(alpha=focal_alpha, gamma=focal_gamma)
-
-        # Avoid double-counting class priors: if alpha != 0.5 (class-balanced),
-        # do NOT pass pos_weight to BCE term. Alpha<0.5 down-weights positives.
-        alpha_diff = abs(float(focal_alpha) - 0.5)
-        pass_pos_weight = alpha_diff < 1e-6
-        if not pass_pos_weight:
-            if focal_alpha < 0.5:
-                logger.warning("focal_alpha < 0.5 down-weights positives; ensure this is intended")
-            logger.info(
-                "[INIT] FOCAL: alpha != 0.5 → disabling pos_weight to avoid double-counting"
-            )
-
-        def compute_loss(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-            if warmup_schedule and warmup_schedule.enabled:
-                effective_gamma = get_focal_gamma(
-                    current_step_ref["value"],
-                    warmup_schedule,
-                    focal_gamma,
-                )
-                if focal.gamma != effective_gamma:
-                    focal.gamma = effective_gamma
-            pw = pos_weight_t if pass_pos_weight else None
-            return cast(torch.Tensor, focal(x, y, pos_weight=pw))
-
-        logger.info(
-            f"[INIT] Using FOCAL loss (alpha={focal_alpha}, gamma={focal_gamma}, "
-            f"pos_weight={'on' if pass_pos_weight else 'off'})"
-        )
-
-        # v3.4.1: Warmup schedules for gradient stabilization
-        if warmup_schedule and warmup_schedule.focal_gamma_enabled:
-            logger.info(
-                f"[WARMUP] Focal gamma warmup enabled: {warmup_schedule.focal_gamma_start:.2f} → "
-                f"{warmup_schedule.focal_gamma_end:.2f} over {warmup_schedule.warmup_steps} steps"
-            )
-    else:
-        bce = nn.BCEWithLogitsLoss(reduction="none", pos_weight=pos_weight_t)
-
-        def compute_loss(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-            return cast(torch.Tensor, bce(x, y))
-
-        logger.info("[INIT] Using BCEWithLogits loss")
-
-    # GUARDRAILS: Validate critical components before training
-    if compute_loss is None or not callable(compute_loss):
-        raise TypeError("compute_loss must be a callable")
-    if scheduler is not None and not hasattr(scheduler, "step"):
-        raise TypeError(f"scheduler must have 'step' method, got {type(scheduler)}")
-    if use_amp and not hasattr(scaler, "scale"):
-        raise TypeError(f"GradScaler expected for AMP, got {type(scaler)}")
-
-    # PREFLIGHT CHECK: Test one batch to catch errors early
-    logger.info("[PREFLIGHT] Testing one batch before training...")
-    test_windows = first_batch["window"].to(device_obj)
-    test_labels = first_batch["label"].to(device_obj)
-
-    # Handle multi-channel labels for test
-    if test_labels.dim() == 3:  # (B, C, T)
-        test_labels = test_labels.max(dim=1)[0]  # (B, T)
-
-    model.eval()
-    try:
-        with torch.no_grad(), autocast(device, enabled=(use_amp and device == "cuda")):
-            if supports_training_state:
-                model.set_training_state(global_step, warmup_schedule)
-            test_logits = model(test_windows)  # (B, T) raw logits
-            test_loss = compute_loss(test_logits, test_labels)
-            if test_loss is None:
-                raise ValueError("Loss computation returned None")
-            logger.info(f"[PREFLIGHT] ✓ Model forward pass OK, loss shape: {test_loss.shape}")
-    except Exception as e:
-        logger.info(f"[PREFLIGHT] ✗ Failed on test batch: {e}")
-        logger.info("[PREFLIGHT] Debug info:")
-        logger.info(f"  - Model type: {type(model)}")
-        logger.info(f"  - Input shape: {test_windows.shape}")
-        logger.info(f"  - Labels shape: {test_labels.shape}")
-        logger.info(f"  - Loss mode: {loss_mode}")
-        logger.info(f"  - Device: {device_obj}")
-        raise
-    finally:
-        model.train()
-
-    logger.info(f"[TRAIN] Starting epoch with {len(dataloader)} batches")
-
-    # v3.4.1: Initialize warmup schedules if enabled
-    if warmup_schedule and warmup_schedule.enabled:
-        model.set_training_state(global_step, warmup_schedule)
-        if warmup_schedule.adj_temperature_enabled:
-            logger.info(
-                f"[WARMUP] Adjacency temperature warmup enabled: "
-                f"{warmup_schedule.adj_temperature_start:.2f} → {warmup_schedule.adj_temperature_end:.2f} "
-                f"over {warmup_schedule.warmup_steps} steps"
-            )
-
-    total_loss = 0.0
-    num_batches = 0
-    consecutive_nans = 0
-    max_consecutive_nans = 50  # Threshold for early termination
-    enable_nan_debug = env.nan_debug()
-    nan_debug_emitted = 0
-    max_nan_debug = env.nan_debug_max()
-
-    from collections import deque
-
-    gradient_norms: deque[float] = deque(maxlen=100)
-
-    # Robust tqdm handling for Modal/non-TTY environments
-    use_tqdm = not env.disable_tqdm()
-    progress_bar = None  # Initialize to None for cleanup
-
-    if use_tqdm:
-        try:
-            # Force tqdm to use simple ASCII output for subprocess compatibility
-            # Disable dynamic ncols to prevent terminal detection issues
-            progress_bar = tqdm(
-                dataloader,
-                desc="Training",
-                leave=False,
-                file=sys.stderr,  # Explicitly use stderr
-                ascii=True,  # Use ASCII characters only
-                ncols=80,  # Fixed width to avoid terminal detection
-                disable=None,  # Let tqdm auto-detect if it should disable
-            )
-            if progress_bar is None or not hasattr(progress_bar, "__iter__"):
-                logger.info("[WARNING] tqdm initialization failed, using plain iteration")
-                progress = dataloader
-            else:
-                progress = progress_bar
-        except Exception as e:
-            logger.info(f"[WARNING] tqdm failed ({e}), using plain iteration")
-            progress = dataloader
-    else:
-        progress = dataloader
-
-    # Use enumerate for batch indexing (satisfies ruff SIM113)
-    # But track global_step separately for proper scheduler behavior
-    try:
-        for batch_idx, batch in enumerate(progress):
-            # CRITICAL: Log batch start BEFORE any processing (detect hangs)
-            if batch_idx == 0 or (batch_idx > 0 and batch_idx % LOG_EVERY_N_STEPS == 0):
-                logger.info(f"[BATCH START] Processing batch {batch_idx}/{len(dataloader)}")
-
-            # Unpack dict format
-            windows = batch["window"].to(device_obj)
-            labels = batch["label"].to(device_obj)
-            # Metadata available but not used during training:
-            # file_ids = batch["file_id"]
-            # window_starts = batch["window_start_s"]
-
-            # Handle multi-channel labels: aggregate to any-seizure
-            if labels.dim() == 3:  # (B, C, T)
-                labels = labels.max(dim=1)[0]  # (B, T)
-
-            # Optional sanitation for non-finite inputs/labels
-            if env.sanitize_inputs():
-                if not torch.isfinite(windows).all():
-                    windows = torch.nan_to_num(windows, nan=0.0, posinf=0.0, neginf=0.0)
-                if not torch.isfinite(labels).all():
-                    labels = torch.nan_to_num(labels, nan=0.0, posinf=0.0, neginf=0.0)
-
-            # Clamp labels to [0,1] for numerical safety
-            if (labels.min() < 0) or (labels.max() > 1):
-                labels = labels.clamp_(0.0, 1.0)
-
-            if supports_training_state:
-                model.set_training_state(global_step, warmup_schedule)
-            current_step_ref["value"] = global_step
-
-            optimizer.zero_grad(set_to_none=True)
-
-            # v3.4.1: Update warmup schedules before forward pass
-            if warmup_schedule and warmup_schedule.enabled:
-                model.set_training_state(global_step, warmup_schedule)
-                # Update focal gamma dynamically if enabled
-                if use_focal and warmup_schedule.focal_gamma_enabled:
-                    effective_gamma = get_focal_gamma(global_step, warmup_schedule, focal_gamma)
-                    focal.gamma = effective_gamma
-
-            # Forward pass with AMP (model returns raw logits)
-            with autocast(device, enabled=(use_amp and device == "cuda")):
-                try:
-                    logits = model(windows)  # (B, T) raw logits
-                    if logits is None:
-                        raise ValueError(f"Model returned None for input shape {windows.shape}")
-                    # Check logits finiteness and sanitize if needed
-                    if not torch.isfinite(logits).all():
-                        nonfinite = (~torch.isfinite(logits)).sum().item()
-                        if batch_idx % LOG_EVERY_N_STEPS == 0:
-                            logger.warning(
-                                f"Non-finite logits at batch {batch_idx}: count={nonfinite} -> sanitizing"
-                            )
-                        # Save bad batch for debugging (best-effort)
-                        try:
-                            Path("debug").mkdir(parents=True, exist_ok=True)
-                            torch.save(
-                                {
-                                    "windows": windows.cpu(),
-                                    "labels": labels.cpu(),
-                                    "global_step": global_step,
-                                },
-                                f"debug/bad_batch_{global_step:06d}.pt",
-                            )
-                        except Exception:
-                            pass
-                        # Sanitize logits to allow training to continue
-                        logits = torch.nan_to_num(logits, nan=0.0, posinf=20.0, neginf=-20.0)
-
-                    # Log forward pass completion for batch 0 (critical for hang detection)
-                    if batch_idx == 0:
-                        logger.info(
-                            f"[BATCH 0] Forward pass completed, logits shape: {logits.shape}"
-                        )
-
-                    per_element_loss = compute_loss(logits, labels)
-                    if per_element_loss is None:
-                        raise ValueError("Loss computation returned None")
-                except Exception as e:
-                    logger.error(f"Forward/loss computation failed at batch {batch_idx}:")
-                    logger.info(f"  - Error: {e}")
-                    logger.info(f"  - Model: {type(model)}")
-                    logger.info(f"  - Windows shape: {windows.shape}")
-                    logger.info(f"  - Labels shape: {labels.shape}")
-                    logger.info(f"  - Device: {windows.device}")
-                    if enable_nan_debug and nan_debug_emitted < max_nan_debug:
-                        try:
-                            w_min = float(windows.min().item())
-                            w_max = float(windows.max().item())
-                            w_mean = float(windows.mean().item())
-                            w_std = float(windows.std().item())
-                            l_min = float(labels.min().item())
-                            l_max = float(labels.max().item())
-                            logger.debug(
-                                f"  - Windows stats: min={w_min:.3e} max={w_max:.3e} mean={w_mean:.3e} std={w_std:.3e}"
-                            )
-                            logger.info(f"  - Labels stats: min={l_min:.3e} max={l_max:.3e}")
-                        except Exception:
-                            pass
-                        nan_debug_emitted += 1
-                    raise
-                # Mean reduction since pos_weight is already in criterion
-                loss = per_element_loss.mean()
-
-            # Check for non-finite loss before gradient update
-            if not torch.isfinite(loss):
-                consecutive_nans += 1
-                if batch_idx % LOG_EVERY_N_STEPS == 0:
-                    logger.warning(
-                        f"NaN loss detected at batch {batch_idx} "
-                        f"(consecutive: {consecutive_nans}), skipping gradient update"
-                    )
-                if enable_nan_debug and nan_debug_emitted < max_nan_debug:
-                    try:
-                        with torch.no_grad():
-                            # Recompute logits in full precision for diagnostics
-                            logits_fp32 = model(windows.float())
-                            nonfinite = (~torch.isfinite(logits_fp32)).sum().item()
-                            logger.debug(
-                                f"FP32 logits non-finite count at batch {batch_idx}: {nonfinite}"
-                            )
-                            # Check batch composition
-                            pos_ratio = labels.sum().item() / labels.numel()
-                            logger.debug(f"Batch {batch_idx} positive ratio: {pos_ratio:.4f}")
-                            # Check for dead channels
-                            channel_stds = windows.std(dim=[0, 2])  # std across batch and time
-                            dead_channels = (channel_stds < 1e-6).sum().item()
-                            if dead_channels > 0:
-                                logger.debug(f"Batch {batch_idx} has {dead_channels} dead channels")
-                    except Exception as e:
-                        logger.info(f"[DEBUG] Error in NaN diagnostics: {e}")
-                    nan_debug_emitted += 1
-                # Clear gradients but skip update
-                optimizer.zero_grad(set_to_none=True)
-
-                # Check if we should stop training
-                if consecutive_nans >= max_consecutive_nans:
-                    logger.critical(
-                        f"{consecutive_nans} consecutive NaN losses detected, "
-                        "model may be corrupted. Stopping training."
-                    )
-                    break
-            else:
-                consecutive_nans = 0  # Reset counter on valid loss
-                # Backward pass with proper scaler handling
-                if scaler.is_enabled():
-                    scaler.scale(loss).backward()
-                    scaler.unscale_(optimizer)
-
-                    # Sanitize gradients if needed
-                    skip_step = False
-                    if env.sanitize_grads():
-                        grad_has_nan = False
-                        for _name, param in model.named_parameters():
-                            if param.grad is not None and not torch.isfinite(param.grad).all():
-                                grad_has_nan = True
-                                param.grad.nan_to_num_(nan=0.0, posinf=0.0, neginf=0.0)
-                        if grad_has_nan:
-                            if batch_idx % LOG_EVERY_N_STEPS == 0:
-                                logger.warning(f"Sanitized NaN gradients at batch {batch_idx}")
-                            if env.skip_opt_step_on_nan():
-                                skip_step = True
-                                logger.warning("Skipping optimizer step due to NaN gradients")
-
-                    if not skip_step:
-                        if gradient_clip > 0:
-                            grad_norm = torch.nn.utils.clip_grad_norm_(
-                                model.parameters(), gradient_clip
-                            )
-                            gradient_norms.append(float(grad_norm))
-                            if enable_nan_debug and grad_norm > gradient_clip * 10:
-                                logger.debug(
-                                    f"Large grad norm at batch {batch_idx}: {grad_norm:.2e} (clipped to {gradient_clip})"
-                                )
-                        scaler.step(optimizer)
-                        scaler.update()
-                else:
-                    loss.backward()
-
-                    # Sanitize gradients if needed
-                    skip_step = False
-                    if env.sanitize_grads():
-                        grad_has_nan = False
-                        for _name, param in model.named_parameters():
-                            if param.grad is not None and not torch.isfinite(param.grad).all():
-                                grad_has_nan = True
-                                param.grad.nan_to_num_(nan=0.0, posinf=0.0, neginf=0.0)
-                        if grad_has_nan:
-                            if batch_idx % LOG_EVERY_N_STEPS == 0:
-                                logger.warning(f"Sanitized NaN gradients at batch {batch_idx}")
-                            if env.skip_opt_step_on_nan():
-                                skip_step = True
-                                logger.warning("Skipping optimizer step due to NaN gradients")
-
-                    if not skip_step:
-                        if gradient_clip > 0:
-                            grad_norm = torch.nn.utils.clip_grad_norm_(
-                                model.parameters(), gradient_clip
-                            )
-                            gradient_norms.append(float(grad_norm))
-                            if enable_nan_debug and grad_norm > gradient_clip * 10:
-                                logger.debug(
-                                    f"Large grad norm at batch {batch_idx}: {grad_norm:.2e} (clipped to {gradient_clip})"
-                                )
-                        optimizer.step()
-
-                # Increment global step counter only after successful update
-                global_step += 1
-
-                # Scheduler step ONLY after a real optimizer step
-                if scheduler is not None:
-                    step_count = getattr(optimizer, "_step_count", 0)
-                    if isinstance(step_count, int) and step_count > 0:
-                        scheduler.step()
-
-            # Handle NaN losses properly
-            loss_val = loss.item()
-            if torch.isfinite(torch.tensor(loss_val)):
-                total_loss += loss_val
-                num_batches += 1
-            else:
-                if batch_idx % LOG_EVERY_N_STEPS == 0:
-                    logger.warning(
-                        f"Non-finite loss detected at batch {batch_idx}, skipping in average"
-                    )
-
-            if use_tqdm and hasattr(progress, "set_postfix"):
-                if not torch.isfinite(torch.tensor(loss_val)):
-                    progress.set_postfix({"loss": "NaN"})
-                else:
-                    progress.set_postfix({"loss": f"{loss_val:.4f}"})
-
-            # Modal progress logging - log based on BGB_LOG_EVERY_N_STEPS
-            if batch_idx > 0 and batch_idx % LOG_EVERY_N_STEPS == 0:
-                current_lr = optimizer.param_groups[0]["lr"]
-                if not torch.isfinite(torch.tensor(loss_val)):
-                    logger.info(
-                        f"[PROGRESS] Batch {batch_idx}/{len(dataloader)} | "
-                        f"Loss: nan | LR: {current_lr:.2e}"
-                    )
-                else:
-                    logger.info(
-                        f"[PROGRESS] Batch {batch_idx}/{len(dataloader)} | "
-                        f"Loss: {loss_val:.4f} | LR: {current_lr:.2e}"
-                    )
-
-            # Heartbeat for Modal (every 5 minutes)
-            if time.time() - last_heartbeat > heartbeat_interval:
-                if num_batches > 0:
-                    avg_loss = total_loss / num_batches
-                    logger.info(
-                        f"[HEARTBEAT] Still training... Batch {batch_idx}/{len(dataloader)} | "
-                        f"Avg Loss: {avg_loss:.4f}"
-                    )
-                else:
-                    logger.info(
-                        f"[HEARTBEAT] Still training... Batch {batch_idx}/{len(dataloader)} | "
-                        f"Avg Loss: N/A (all NaN)"
-                    )
-                last_heartbeat = time.time()
-
-                mem_stats = get_memory_stats()
-                if mem_stats:
-                    mem_log = "[MEMORY]"
-                    if "gpu_allocated_gb" in mem_stats:
-                        mem_log += f" GPU: {mem_stats['gpu_allocated_gb']:.2f}GB alloc"
-                        mem_log += f" / {mem_stats['gpu_reserved_gb']:.2f}GB res"
-                    if "ram_used_gb" in mem_stats:
-                        mem_log += f" | RAM: {mem_stats['ram_used_gb']:.2f}GB used"
-                        mem_log += f" / {mem_stats['ram_available_gb']:.2f}GB avail"
-                    if "swap_used_gb" in mem_stats and mem_stats["swap_used_gb"] > 0.1:
-                        mem_log += f" | SWAP: {mem_stats['swap_used_gb']:.2f}GB"
-                    logger.info(mem_log)
-
-                if len(gradient_norms) > 10:
-                    sorted_norms = sorted(gradient_norms)
-                    n = len(sorted_norms)
-                    grad_mean = sum(sorted_norms) / n
-                    grad_p50 = sorted_norms[n // 2]
-                    grad_p95 = sorted_norms[int(n * 0.95)]
-                    grad_max = sorted_norms[-1]
-                    logger.info(
-                        f"[GRADIENTS] Last {n} batches: "
-                        f"Mean={grad_mean:.2f} | P50={grad_p50:.2f} | "
-                        f"P95={grad_p95:.2f} | Max={grad_max:.2f}"
-                    )
-
-            if (
-                checkpoint_dir is not None
-                and epoch_index is not None
-                and mid_interval_s is not None
-                and (time.time() - last_mid_save) >= mid_interval_s
-            ):
-                mid_path = checkpoint_dir / f"mid_epoch_{epoch_index + 1:03d}_{batch_idx:06d}.pt"
-                try:
-                    save_checkpoint(
-                        model,
-                        optimizer,
-                        epoch_index,
-                        0.0,
-                        mid_path,
-                        scheduler,
-                        None,
-                        extra={"batch_idx": batch_idx, "kind": "mid_epoch"},
-                    )
-                    logger.info(f"[CHECKPOINT] Saved mid-epoch snapshot: {mid_path.name}")
-                    last_mid_save = time.time()
-                    mids = sorted(
-                        checkpoint_dir.glob("mid_epoch_*.pt"), key=lambda p: p.stat().st_mtime
-                    )
-                    if len(mids) > int(max(0, mid_epoch_keep)):
-                        for old in mids[: len(mids) - int(mid_epoch_keep)]:
-                            with suppress(Exception):
-                                old.unlink()
-                except Exception as e:
-                    logger.info(f"[WARNING] Failed to save mid-epoch checkpoint: {e}")
-
-    except Exception as e:
-        # Clean up tqdm if it exists
-        if progress_bar is not None and hasattr(progress_bar, "close"):
-            with suppress(Exception):
-                progress_bar.close()
-        # Re-raise the actual error with context
-        logger.info(f"[ERROR] Training loop failed at batch {num_batches}: {e}")
-        raise
-    finally:
-        # Always clean up tqdm progress bar
-        if progress_bar is not None and hasattr(progress_bar, "close"):
-            with suppress(Exception):
-                progress_bar.close()
-
-    avg_loss = total_loss / max(1, num_batches)
-    return (avg_loss, global_step) if return_step else avg_loss
-
-
 # ============================================================================
 # Validation epoch (Single Responsibility)
 # ============================================================================
